Reorder e2e test suites into more logical categorization (#6036)
* WIP

* refactor exploration editor/player test suites

* linting fixes

* remove extra file

* split exploration tabs into separate test files

* change configuration file to remove old files, and include new ones

* alphabetize, and changes to .travis.yml

* change year to 2019

* fix missing imports in test suites

* fix stray space

* disable travis test for blank test file

* delete unused suite
