+++ core/tests/protractor_desktop/userJourneys.js
@@ -26,7 +26,6 @@ var ExplorationPlayerPage = require(
   '../protractor_utils/ExplorationPlayerPage.js');
 var LibraryPage = require('../protractor_utils/LibraryPage.js');
 var PreferencesPage = require('../protractor_utils/PreferencesPage.js');
-var ThanksPage = require('../protractor_utils/ThanksPage.js');
 var forms = require('../protractor_utils/forms.js');
 var general = require('../protractor_utils/general.js');
 var users = require('../protractor_utils/users.js');
@@ -84,131 +83,7 @@ describe('Basic user journeys', function() {
       general.checkForConsoleErrors([]);
     });
   });
-
-  describe('Preferences', function() {
-    var preferencesPage = null;
-
-    beforeEach(function() {
-      preferencesPage = new PreferencesPage.PreferencesPage();
-    });
-
-    it('should change editor role email checkbox value', function() {
-      users.createUser('alice@preferences.com', 'alicePreferences');
-      users.login('alice@preferences.com');
-      preferencesPage.get();
-      expect(preferencesPage.isEditorRoleEmailsCheckboxSelected()).toBe(true);
-      preferencesPage.toggleEditorRoleEmailsCheckbox();
-      expect(preferencesPage.isEditorRoleEmailsCheckboxSelected()).toBe(false);
-      browser.refresh();
-      expect(preferencesPage.isEditorRoleEmailsCheckboxSelected()).toBe(false);
-    });
-
-    it('should change feedback message email checkbox value', function() {
-      users.createUser('bob@preferences.com', 'bobPreferences');
-      users.login('bob@preferences.com');
-      preferencesPage.get();
-      expect(preferencesPage.isFeedbackEmailsCheckboxSelected()).toBe(true);
-      preferencesPage.toggleFeedbackEmailsCheckbox();
-      expect(preferencesPage.isFeedbackEmailsCheckboxSelected()).toBe(false);
-      browser.refresh();
-      expect(preferencesPage.isFeedbackEmailsCheckboxSelected()).toBe(false);
-    });
-
-    afterEach(function() {
-      general.checkForConsoleErrors([]);
-      users.logout();
-    });
-  });
 });
-
-describe('Oppia static pages tour', function() {
-  var thanksPage = null;
-
-  beforeEach(function() {
-    browser.get(general.SERVER_URL_PREFIX);
-    waitFor.pageToFullyLoad();
-  });
-
-  it('visits the links in About dropdown', function() {
-    var LINKS_CLASS_NAMES = [
-      '.protractor-test-about-link',
-      '.protractor-test-get-started-link',
-      '.protractor-test-playbook-link'
-    ];
-
-    LINKS_CLASS_NAMES.forEach(function(className) {
-      var dropdown = element(by.css('.protractor-test-about-oppia-list-item'));
-      browser.actions().mouseMove(dropdown).perform();
-      dropdown.element(by.css(className)).click();
-      waitFor.pageToFullyLoad();
-    });
-  });
-
-  it('visits the donate link', function() {
-    element(by.css('.protractor-test-donate-link')).click();
-    waitFor.pageToFullyLoad();
-  });
-
-  it('visits the thanks for donating page', function() {
-    thanksPage = new ThanksPage.ThanksPage();
-    thanksPage.get();
-  });
-
-  it('visits the terms page', function() {
-    element(by.css('.protractor-test-terms-link')).click();
-    waitFor.pageToFullyLoad();
-  });
-
-  it('visits the privacy page', function() {
-    element(by.css('.protractor-test-privacy-policy-link')).click();
-    waitFor.pageToFullyLoad();
-  });
-
-  it('visits the Fractions landing page', function() {
-    browser.get('/fractions');
-    waitFor.pageToFullyLoad();
-  });
-
-  it('visits the Partners landing page', function() {
-    browser.get('/partners');
-    waitFor.pageToFullyLoad();
-  });
-
-  it('visits the Nonprofits landing page', function() {
-    browser.get('/nonprofits');
-    waitFor.pageToFullyLoad();
-  });
-
-  it('visits the Parents landing page', function() {
-    browser.get('/parents');
-    waitFor.pageToFullyLoad();
-  });
-
-  it('visits the Teachers landing page', function() {
-    browser.get('/teachers');
-    waitFor.pageToFullyLoad();
-  });
-
-  it('visits the Volunteers landing page', function() {
-    browser.get('/volunteers');
-    waitFor.pageToFullyLoad();
-  });
-
-  afterEach(function() {
-    general.checkForConsoleErrors([
-      // TODO (Jacob) Remove when
-      // https://code.google.com/p/google-cast-sdk/issues/detail?id=309 is fixed
-      'cast_sender.js - Failed to load resource: net::ERR_FAILED',
-      'Uncaught ReferenceError: ytcfg is not defined',
-      // TODO (@pranavsid98) This error is caused by the upgrade from Chrome 60
-      // to Chrome 61. Chrome version at time of recording this is 61.0.3163.
-      'chrome-extension://invalid/ - Failed to load resource: net::ERR_FAILED',
-      'Error parsing header X-XSS-Protection: 1; mode=block; ' +
-      'report=https:\/\/www.google.com\/appserve\/security-bugs\/log\/youtube:',
-    ]);
-  });
-});
-
 
 describe('Site language', function() {
   var adminPage = null;
